Move `RawTableInner` and `TableLayout` to separate module
